Rename Lens.Micro.Mtl.Zoom to Lens.Micro.Mtl.Internal and expose it
